implementar testes de integração para os novos endpoints e ajustes gerais

+++ src/main/java/com/williamsilva/avaliacaofilmesapi/domain/repository/projections/AnoContagem.java
@@ -3,5 +3,5 @@ package com.williamsilva.avaliacaofilmesapi.domain.repository.projections;
 public interface AnoContagem {
 
     Integer getAno();
-    Long getContagem();
+    Integer getContagem();
 }
