+++ src/main/java/com/williamsilva/avaliacaofilmesapi/domain/repository/projections/EstudioContagem.java
@@ -1,5 +1,5 @@
 package com.williamsilva.avaliacaofilmesapi.domain.repository.projections;
 
-public record EstudioContagem(String estudio, Long contagem) {
+public record EstudioContagem(String estudio, Integer contagem) {
 
 }
